US4565: Refactor Complete
