+++ crossroads.net/app/formlyBuilder/formlyBuilder.module.js
@@ -2,10 +2,10 @@
   'use strict';
 
   var MODULE = require('crds-constants').MODULES.FORMLY_BUILDER;
-
-  angular.module(MODULE, ['crossroads.core', 
-                          'crossroads.common'
-                          ])
-    .config(require('./formlyBuilder.config'))
-    .run(require('./formlyBuilder.run'));
+  var ngModule = angular.module(MODULE, ['crossroads.core', 'crossroads.common']);
+  
+  ngModule.run(require('./formlyBuilder.validationConfig'));
+  
+  require('./types')(ngModule);
+  require('./wrappers')(ngModule);
 })();
