+++ crossroads.net/app/formlyBuilder/types/datePicker.js
@@ -1,0 +1,82 @@
+export default ngModule => {
+    ngModule.run(addDatePicker);
+
+    function addDatePicker(formlyConfig) {
+        var attributes = [
+            'date-disabled',
+            'custom-class',
+            'show-weeks',
+            'starting-day',
+            'init-date',
+            'min-mode',
+            'max-mode',
+            'format-day',
+            'format-month',
+            'format-year',
+            'format-day-header',
+            'format-day-title',
+            'format-month-title',
+            'year-range',
+            'shortcut-propagation',
+            'datepicker-popup',
+            'show-button-bar',
+            'current-text',
+            'clear-text',
+            'close-text',
+            'close-on-date-selection',
+            'datepicker-append-to-body'
+        ];
+
+        var bindings = [
+            'datepicker-mode',
+            'min-date',
+            'max-date'
+        ];
+
+        var ngModelAttrs = {};
+
+        angular.forEach(attributes, function (attr) {
+            ngModelAttrs[camelize(attr)] = { attribute: attr };
+        });
+
+        angular.forEach(bindings, function (binding) {
+            ngModelAttrs[camelize(binding)] = { bound: binding };
+        });
+
+        formlyConfig.setType({
+            name: 'datepicker',
+            template: require('./templates/datepicker.html'),
+            wrapper: ['formlyBuilderHasError', 'formlyBuilderLabel'],
+            defaultOptions: {
+                ngModelAttrs: ngModelAttrs,
+                templateOptions: {
+                    datepickerOptions: {
+                        format: 'MM/dd/yyyy',
+                        initDate: new Date()
+                    }
+                }
+            },
+            controller: ['$scope', function ($scope) {
+                $scope.datepicker = {};
+
+                $scope.datepicker.opened = false;
+
+                $scope.datepicker.open = function ($event) {
+                    $scope.datepicker.opened = !$scope.datepicker.opened;
+                };
+            }]
+        });
+
+        //this function is being used in multiple type definitions,
+        //need to pull it out and make it a common tool
+        function camelize(string) {
+            string = string.replace(/[\-_\s]+(.)?/g, function (match, chr) {
+                return chr ? chr.toUpperCase() : '';
+            });
+            // Ensure 1st char is always lowercase
+            return string.replace(/^([A-Z])/, function (match, chr) {
+                return chr ? chr.toLowerCase() : '';
+            });
+        }
+    }
+}
