+++ crossroads.net/app/formlyBuilder/types/formlyBuilderInput.js
@@ -1,0 +1,12 @@
+export default ngModule => {
+    ngModule.config(addFormlyBuilderCheckbox);
+
+    function addFormlyBuilderCheckbox(formlyConfigProvider) {
+        formlyConfigProvider.setType({
+            name: 'formlyBuilderInput',
+            template: '<input class="form-control" ng-model="model[options.key]">',
+            wrapper: ['formlyBuilderHasError', 'formlyBuilderLabel']
+        });
+    }
+}
+
