+++ crossroads.net/app/formlyBuilder/types/index.js
@@ -1,0 +1,15 @@
+export default ngModule => {
+  require('./multiCheckBoxCombo')(ngModule);
+  require('./zipCodeValidator')(ngModule);
+  require('./boldCheckBox')(ngModule);
+  require('./profilePicture')(ngModule);
+  require('./timePicker')(ngModule);
+  require('./datePicker')(ngModule);
+  require('./formlyBuilderInput')(ngModule);
+  require('./formlyBuilderCheckbox')(ngModule);
+  require('./formlyBuilderMultiCheckbox')(ngModule);
+  require('./formlyBuilderRadio')(ngModule);
+  require('./formlyBuilderRadioDesc')(ngModule);
+  require('./formlyBuilderSelect')(ngModule);
+  require('./formlyBuilderTextarea')(ngModule);
+}
