+++ crossroads.net/app/formlyBuilder/types/templates/boldCheckbox.html
