+++ crossroads.net/app/formlyBuilder/types/templates/datepicker.html
