+++ crossroads.net/app/formlyBuilder/types/templates/formlyBuilder-checkbox.html
