+++ crossroads.net/app/formlyBuilder/types/templates/multiCheckBoxCombo.html
