+++ crossroads.net/app/formlyBuilder/wrappers/formlyBuilder-has-error.js
@@ -1,0 +1,10 @@
+export default ngModule => {
+    ngModule.config(addFormlyBuilderHasError);
+
+    function addFormlyBuilderHasError(formlyConfigProvider) {
+        formlyConfigProvider.setWrapper({
+            name: 'formlyBuilderHasError', 
+            template: require('./templates/formlyBuilder-has-error.html')
+        });
+    }
+}
