+++ crossroads.net/app/formlyBuilder/wrappers/formlyBuilder-show-error.js
@@ -1,0 +1,10 @@
+export default ngModule => {
+    ngModule.config(addFormlyBuilderShowError);
+
+    function addFormlyBuilderShowError(formlyConfigProvider) {
+        formlyConfigProvider.setWrapper({
+            template: require('./templates/formlyBuilder-show-error.html')
+        });
+    }
+}
+
