+++ crossroads.net/app/formlyBuilder/wrappers/index.js
@@ -1,0 +1,5 @@
+export default ngModule => {
+  require('./formlyBuilder-show-error')(ngModule);
+  require('./formlyBuilder-label')(ngModule);
+  require('./formlyBuilder-has-error')(ngModule);
+}
