+++ crossroads.net/app/formlyBuilder/wrappers/templates/formlyBuilder-show-error.html
@@ -1,0 +1,6 @@
+<formly-transclude></formly-transclude>
+<div class="validation" ng-if="showError" ng-messages="fc.$error || fc[0].$error">
+    <div ng-message="{{::name}}" ng-repeat="(name, message) in ::options.validation.messages">
+        {{message(fc.$viewValue, fc.$modelValue, this)}}
+    </div>
+</div>
